+++ monks-common-display-layer.js
@@ -1,0 +1,15 @@
+export class MonksCommonDisplayLayer extends PlaceablesLayer {
+    constructor() {
+        super();
+    }
+
+    static get layerOptions() {
+        return mergeObject(super.layerOptions, {
+            objectClass: Note,
+            sheetClass: NoteConfig,
+            canDragCreate: false,
+            zIndex: 180
+        });
+    }
+
+}