+++ settings.js
@@ -2,8 +2,18 @@ import { MonksCommonDisplay, i18n } from "./monks-common-display.js";
 
 export const registerSettings = function () {
     // Register any custom module settings here
-    let modulename = "monks-common-display";
+	let modulename = "monks-common-display";
 
+	game.settings.registerMenu(modulename, 'hot-keys', {
+		name: 'Change Hotkeys',
+		label: 'Change Hotkeys',
+		hint: 'Change the hotkeys that this module uses',
+		icon: 'fas fa-keyboard',
+		restricted: true,
+		type: Hotkeys.createConfig('Monks Common Display', ['monks-common-display'])
+	});
+
+	/*
 	game.settings.register(modulename, "display-players", {
 		name: i18n("MonksCommonDisplay.display-players.name"),
 		hint: i18n("MonksCommonDisplay.display-players.hint"),
@@ -11,7 +21,7 @@ export const registerSettings = function () {
 		config: true,
 		default: "",
 		type: String,
-	});
+	});*/
 	game.settings.register(modulename, "mirror-movement", {
 		name: i18n("MonksCommonDisplay.mirror-movement.name"),
 		hint: i18n("MonksCommonDisplay.mirror-movement.hint"),
@@ -20,6 +30,7 @@ export const registerSettings = function () {
 		default: true,
 		type: Boolean,
 	});
+	/*
 	game.settings.register(modulename, "show-mirror-tool", {
 		name: i18n("MonksCommonDisplay.show-mirror-tool.name"),
 		hint: i18n("MonksCommonDisplay.show-mirror-tool.hint"),
@@ -27,7 +38,7 @@ export const registerSettings = function () {
 		config: true,
 		default: true,
 		type: Boolean,
-	});
+	});*/
 	game.settings.register(modulename, "show-chat-log", {
 		name: i18n("MonksCommonDisplay.show-chat-log.name"),
 		hint: i18n("MonksCommonDisplay.show-chat-log.hint"),
@@ -43,5 +54,23 @@ export const registerSettings = function () {
 		config: true,
 		default: true,
 		type: Boolean,
+	});
+
+	game.settings.register(modulename, "startupdata", {
+		name: '',
+		hint: '',
+		scope: "client",
+		config: false,
+		default: false,
+		type: Boolean,
+	});
+
+	game.settings.register(modulename, "playerdata", {
+		name: '',
+		hint: '',
+		scope: "world",
+		config: false,
+		default: {},
+		type: Object,
 	});
 };