+++ CHANGELOG.md
@@ -1,1 +1,4 @@
+# Version 1.0.1
+Added hotkey support for changing Mirror Screen setting
 
+Adding support for Image Hover module